+++ tests/tests_integration/test_alerts/test_alerts.py
@@ -1,9 +1,10 @@
 import os
 import random
 from datetime import datetime, timezone
-from time import time
-from typing import Callable
+from typing import Callable, Union
 
+from cognite.client import ClientConfig
+from cognite.client.credentials import OAuthClientCredentials
 from pytest import fixture, mark
 
 from cognite.experimental import CogniteClient
@@ -11,9 +12,12 @@ from cognite.experimental.data_classes.alerts import (
     Alert,
     AlertChannel,
     AlertChannelUpdate,
+    AlertList,
     AlertSubscriber,
     AlertSubscription,
     AlertSubscriptionDelete,
+    DeduplicateAlert,
+    DeduplicateAlertList,
 )
 
 CURRENT_TS = datetime.now(timezone.utc)
@@ -24,7 +28,24 @@ ALERTS_INT_TEST_EMAIL = f"ivan.polomanyi+{CURRENT_TS_INT}@cognite.com"
 
 @fixture(scope="class")
 def cognite_client() -> CogniteClient:
-    return CogniteClient(headers={"cdf-version": "alpha"})
+    creds = OAuthClientCredentials(
+        token_url=os.environ["COGNITE_TOKEN_URL"],
+        client_id=os.environ["COGNITE_CLIENT_ID"],
+        client_secret=os.environ["COGNITE_CLIENT_SECRET"],
+        scopes=[os.environ["COGNITE_TOKEN_SCOPES"]],
+    )
+    return CogniteClient(
+        config=ClientConfig(
+            base_url=os.environ["COGNITE_BASE_URL"]
+            if os.environ["COGNITE_BASE_URL"]
+            else "https://azure-dev.cognitedata.com",
+            client_name="experimental",
+            project="air-azure-dev",
+            headers={"cdf-version": "alpha"},
+            credentials=creds,
+            debug=True,
+        )
+    )
 
 
 @fixture(scope="class")
@@ -44,6 +65,28 @@ def base_alert() -> Callable[..., Alert]:
 
 
 @fixture(scope="class")
+def base_alert_with_triggered_points() -> Callable[..., Alert]:
+    def create(channel_external_id: int) -> Alert:
+        return Alert(
+            # external_id="test_" + CURRENT_TS_STR + str(random.random()),
+            channel_external_id=channel_external_id,
+            source="inttest",
+            level="INFO",
+            metadata={"test": "test"},
+            value="10 percent",
+            triggered_points=[
+                {"triggered": True, "timestamp": 1000000},
+                {"triggered": True, "timestamp": 1120000},
+                {"triggered": True, "timestamp": 3500000},
+                {"triggered": False, "timestamp": 4900000},
+                {"triggered": True, "timestamp": 5000000},
+            ],
+        )
+
+    return create
+
+
+@fixture(scope="class")
 def base_channel() -> Callable[..., AlertChannel]:
     def create() -> AlertChannel:
         return AlertChannel(
@@ -51,6 +94,20 @@ def base_channel() -> Callable[..., AlertChannel]:
             name="test_channel_" + CURRENT_TS_STR,
             description="integration test channel",
             metadata={"test": "test"},
+        )
+
+    return create
+
+
+@fixture(scope="class")
+def base_channel_with_deduplication_rules() -> Callable[..., AlertChannel]:
+    def create() -> AlertChannel:
+        return AlertChannel(
+            external_id="test_" + CURRENT_TS_STR + "_" + str(random.random()),
+            name="test_channel_" + CURRENT_TS_STR,
+            description="integration test channel with deduplication rules",
+            metadata={"test": "test"},
+            alert_rules={"deduplication": {"merge_interval": "2m", "activation_interval": "5m"}},
         )
 
     return create
@@ -91,12 +148,14 @@ class TestAlertChannelsIntegration:
 
     def test_create_2(self, cognite_client, base_channel):
         res = cognite_client.alerts.channels.create([base_channel(), base_channel()])
+        assert len(res) == 2
 
+    def test_create_3(self, cognite_client, base_channel, base_channel_with_deduplication_rules):
+        res = cognite_client.alerts.channels.create([base_channel(), base_channel_with_deduplication_rules()])
         assert len(res) == 2
 
     def test_list(self, cognite_client):
         res = cognite_client.alerts.channels.list()
-
         assert len(res) > 0
 
     def test_update_by_alert_channel(self, cognite_client, base_channel):
@@ -138,8 +197,17 @@ class TestAlertChannelsIntegration:
         assert updated[0].metadata == {"a": "b", "test": "test"}
         assert "updated_ext_id" in updated[0].external_id
 
-    def test_delete(self, cognite_client, base_channel):
+    def test_delete_channel(self, cognite_client, base_channel):
         created = cognite_client.alerts.channels.create(base_channel())
+
+        cognite_client.alerts.channels.delete(external_ids=[created.external_id])
+
+        res = cognite_client.alerts.channels.list(external_ids=[created.external_id])
+
+        assert len(res) == 0
+
+    def test_delete_channel_with_deduplication_rules(self, cognite_client, base_channel_with_deduplication_rules):
+        created = cognite_client.alerts.channels.create(base_channel_with_deduplication_rules())
 
         cognite_client.alerts.channels.delete(external_ids=[created.external_id])
 
@@ -165,6 +233,13 @@ class TestAlertsIntegration:
         res = cognite_client.alerts.create([base_alert(channel.external_id), base_alert(channel.external_id)])
 
         assert len(res) == 2
+
+    def test_create_3(self, cognite_client, base_alert_with_triggered_points, base_channel_with_deduplication_rules):
+        channel = cognite_client.alerts.channels.create(base_channel_with_deduplication_rules())
+
+        res = cognite_client.alerts.create_deduplicated(base_alert_with_triggered_points(channel.external_id))
+        assert isinstance(res, Union[Alert, AlertList])
+        assert channel.external_id == res.channel_external_id
 
     def test_close(self, cognite_client):
         alerts = cognite_client.alerts.list(closed=False)